requirements.txt
langchain
langchain-openai
langchain-pinecone
langchain-community
python-dotenv
pandas
ipython
streamlit